backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
sqlalchemy
psycopg2-binary
passlib
bcrypt==4.0.1
python-jose[cryptography]
